pyproject.toml
[project]
name = "duckmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.20.0",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
]
